change long to size_t or int

+++ cocos/2d/CCActionManager.h
@@ -102,10 +102,10 @@ public:
      * - If you are running 1 Sequence of 7 actions, it will return 1.
      * - If you are running 7 Sequences of 2 actions, it will return 7.
      */
-    long getNumberOfRunningActionsInTarget(const Object *target) const;
+    size_t getNumberOfRunningActionsInTarget(const Object *target) const;
 
     /** @deprecated use getNumberOfRunningActionsInTarget() instead */
-    CC_DEPRECATED_ATTRIBUTE inline unsigned int numberOfRunningActionsInTarget(Object *target) const { return getNumberOfRunningActionsInTarget(target); }
+    CC_DEPRECATED_ATTRIBUTE inline size_t numberOfRunningActionsInTarget(Object *target) const { return getNumberOfRunningActionsInTarget(target); }
 
     /** Pauses the target: all running actions and newly added actions will be paused.
     */
