+++ cocos/2d/platform/CCFileUtils.h
@@ -88,7 +88,7 @@ public:
      *  @return Upon success, a pointer to the data is returned, otherwise NULL.
      *  @warning Recall: you are responsible for calling free() on any Non-NULL pointer returned.
      */
-    virtual unsigned char* getFileData(const char* filename, const char* mode, long *size);
+    virtual unsigned char* getFileData(const char* filename, const char* mode, size_t *size);
 
     /**
      *  Gets resource file data from a zip file.
